more descriptive LoadError for spork/ext/ruby-debug
if ruby-debug not installed and the project loads spork/ext/ruby-debug, then it shows an error saying 'ruby-debug' not installed.

This can confuse the user, as they required 'spork/ext/ruby-debug' in the project, not ruby-debug.  Why didn't spork install ruby-debug for me?

Well, the answer is because spork/ext/ruby-debug is optional.  This new error should make that clear.