Reapply the rails inherited magic for controllers
There is way too much pushing rails around already to achieve what rails is setting out to achieve. Hopefully interfaces and magic is cleaned up in rails 3.0 so that this is less difficult.

When a class inherits from ApplicationController, it gets a brand new master_helper_module, which then copies all the helper methods from it's parent (ApplicationController).  Since at that point, what it was pointing to was the stubbed ApplicationController, it didn't get all the helper methods that would've been in the real ApplicationController. Therefore, reapplying inheritance causes the latest copy of the master_helper_module in ApplicationController to propagate through all subclasses.

Change was made in response to this issue: http://github.com/timcharper/spork/issues#issue/17