pass string to require -- fix for RubyGems bug
SPORK_ROOT is a pathname object and the result of adding 
a string to it is a Pathname object.

RubyGems 1.6.0 through 1.8.5 pass the require path to
Regexp.escape(). In Ruby 1.9.2 passing a Pathname to
Regex.escape() throws this error:

  "can't convert Pathname to String (TypeError)"

This RubyGems bug is fixed but not yet in a release.

This bug does not show up when using bundler exec to run
cucumber feature tests -- only when using cucumber directly.