fix cucumber-rails integration and feature test
See: https://github.com/timcharper/spork/issues/122

I did not commit the Gemfile.lock file used for generated rails3
application used for the cucumber-rails integration feature test.

This file does not affect the operation of spork and
will be re-generated whenever the Gemfile is updated and
the feature is run. 

FYI: the following feature/scenarios fail on the master 
branch both before and after this commit:

  features/diagnostic_mode.feature:6 
  # Scenario: Running spork --diagnose
  
  features/rails_delayed_loading_workarounds.feature:70 
  # Scenario: respecting custom autoload paths